% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{echo}
\alias{echo}
\title{Print to renamer GUI}
\usage{
echo(message, printToGUI = TRUE, printToConsole = TRUE)
}
\arguments{
\item{message}{character, the message to be printed.}

\item{printToGUI}{logical, indicating whether or not to print the message to the gtext widget on the renamer GUI.}

\item{printToConsole}{logical, indicating whether or not to print the message to the console.}
}
\description{
Prints a message to the console and the renamer tool GUI and/or the R console.
When printing to the console, just calls the print function.
}
