% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leica.r
\name{getfiles.XML.LRP}
\alias{getfiles.XML.LRP}
\title{Find Leica MatrixScreener XML and LRP files}
\usage{
getfiles.XML.LRP(folder)
}
\arguments{
\item{folder}{character, Leica MatrixScreener export folder.}
}
\value{
character vector with file names.
}
\description{
Analyzes the Leica MatrixScreener 'AdditionalData' folder and returns the file names of {ScanningTemplate} files
}
\note{
There must be exactly one copy of these files:\cr
\itemize{
  \item\code{\{ScanningTemplate\}templatename.xml}\cr
  \item\code{\{ScanningTemplate\}templatename.lrp}\cr
}
}
