% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{newinfile.char}
\alias{newinfile.char}
\title{Create empty infile (character vector)}
\usage{
newinfile.char(numrow = 8, numcol = 12, show = FALSE,
  saveto = character())
}
\arguments{
\item{numrow}{integer, the number of rows in the multiwell plate. The default value is 8 (i.e. 96 well plate).}

\item{numcol}{integer, the number of columns in the multiwell plate. The default value is 12 (i.e. 96 well plate).}

\item{show}{logical, indicating whether or not to print the result to the console.}

\item{saveto}{character. If specified, the empty infile template is saved in a file at this location.}
}
\value{
Character vector with empty infile template.
}
\description{
Creates an empty microscope infile template, which can be printed to the console and/or saved as a file.
The output is structured as a character vector.
See the \strong{Note} section for a description of the infile structure.
}
\note{
The microscope infile is a text file with the following structure:\cr

    \code{001--A--01--00--00--data1--data2}\cr
    \code{002--A--02--01--00--data1--data2}\cr
    \code{003--A--03--02--00--data1--data2}\cr

In the infile, metadata fields are separated by a double dash (\code{--}). The meaning of the metadata fields is the following:
\itemize{
  \item Well number
  \item Row label (A, B, ...)
  \item Column label (01, 02, 03, ...)
  \item Column label (00, 01, 02, ...)
  \item Row label (00, 01, 02, ...)
  \item Experimental data 1
  \item Experimental data 2
}

Wells are numbered in ascending order, line by line. In the case of a 96 well plate, wells are numbered as shown:

\code{-----1---2---3---4---5---6---7---8---9--10--11--12-}
\code{---------------------------------------------------}
\code{A | 001 002 003 004 005 006 007 008 009 010 011 012}
\code{B | 013 014 015 016 017 018 019 020 021 022 023 024}
\code{C | 025 026 027 028 029 030 031 032 033 034 035 036}
\code{D | 037 038 039 040 041 042 043 044 045 046 047 048}
\code{E | 049 050 051 052 053 054 055 056 057 058 059 060}
\code{F | 061 062 063 064 065 066 067 068 069 070 071 072}
\code{G | 073 074 075 076 077 078 079 080 081 082 083 084}
\code{H | 085 086 087 088 089 090 091 092 093 094 095 096}

The experimental data fields may contain any relevant information describing the well contents.
}
\examples{

# Infile template for a 96 well plate
newinfile.char()

# Save 384 well plate infile template to the working directory.
# The TXT file extension is recommended.
newinfile.char(numrow = 16, numcol = 24, saveto = "./infile_template.txt")

}
\seealso{
Function \code{\link{newinfile.df}} creates an empty infile in the form of a data frame.\cr
Function \code{\link{read.infile.df}} reads infile text files from disk.\cr
}
