% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leica.r
\name{rename_leica}
\alias{rename_leica}
\title{Leica renamer}
\usage{
rename_leica(
  sourcefolder,
  targetfolder,
  infilepath,
  compress = FALSE,
  move = FALSE,
  outputDescriptors = TRUE,
  printMessages = TRUE,
  printFiles = TRUE,
  printToGUI = FALSE
)
}
\arguments{
\item{sourcefolder}{character, the folder with raw images.}

\item{targetfolder}{character, folder where to place renamed files. A subfolder wih the same name as the infile will be created by the function and must not exist previously.}

\item{infilepath}{character, location of the microscope infile}

\item{compress}{logical, compress raw images?}

\item{move}{logical, move (rather than copy) raw images?}

\item{outputDescriptors}{logical, save plain text files with experimental metadata?}

\item{printMessages, }{logical, print messages in the console?}

\item{printFiles, }{logical, print file-by-file progress in the console?}

\item{printToGUI}{logical, print messages to the renamer GUI?}
}
\value{
None, this function just calls another function to perform the renaming.
}
\description{
Renaming of Leica files according to a microscope infile. Auto-detects the software which generated the images to call the apropriate renamer function. See \code{\link{newinfile.df}} or \code{\link{newinfile.char}} for a description of the infile structure.
}
\seealso{
\code{\link{rename_leica_gui}} for interacting with this function using a GUI.
}
