% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leica.r
\name{rename_leica_gui}
\alias{rename_leica_gui}
\title{Leica renamer GUI}
\usage{
rename_leica_gui()
}
\value{
Null
}
\description{
Displays the renamer tool GUI for Leica LAS X MatrixsSreener exports.
The actual renaming process is performed by function rename_leica().
}
\seealso{
\code{\link{rename_leica}} is the function which implements the actual renaming algorithm.\cr
\code{\link{rename_zeiss_gui}} for the Zeiss renamer GUI.\cr
}
