% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olympus.r
\name{rename_olympus}
\alias{rename_olympus}
\title{Olympus renamer}
\usage{
rename_olympus(
  sourcefolder,
  targetfolder,
  infilepath,
  numrow,
  numcol,
  REGEXvalidimages = "^.*(\\\\d{4})_(\\\\d{2}).oib$",
  printMessages = TRUE,
  printFiles = TRUE,
  printToGUI = FALSE,
  move = FALSE
)
}
\arguments{
\item{sourcefolder}{character, the folder with raw images}

\item{targetfolder}{character, folder where to place renamed files. A subfolder with the same name as the infile will be created by the function and must not exist previously.}

\item{infilepath}{character, location of the microscope infile}

\item{numrow}{integer, number of rows in the multi well plate}

\item{numcol}{integer, number of columns in the multi well plate}

\item{REGEXvalidimages}{character, regular expression matching all files which should be renamed. Captures (\code{(\1)}) the scene number (i.e. well number) and (\code{(\2)}) the time number in a time lapse.}

\item{printMessages, }{logical, print messages in the console?}

\item{printFiles, }{logical, print file-by-file progress in the console?}

\item{printToGUI}{logical, print messages to the renamer GUI?}

\item{move}{logical, move (rather than copy) raw images?}
}
\value{
Renamed files, metadata and log file

Null
}
\description{
Renaming of Olympus OIB files according to a microscope infile. See \code{\link{newinfile.df}} or \code{\link{newinfile.char}} for a description of the infile structure.
}
\seealso{
\code{\link{rename_olympus_gui}} for interacting with this function using a GUI.
}
