% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olympus.r
\name{rename_olympus_gui}
\alias{rename_olympus_gui}
\title{Olympus renamer GUI}
\usage{
rename_olympus_gui()
}
\value{
Null
}
\description{
Displays the renamer tool GUI for Olympus OIB files.
The actual renaming process is performed by function rename_olympus().
}
\seealso{
\code{\link{rename_olympus}} is the function which implements the actual renaming algorithm.\cr
\code{\link{rename_leica_gui}} for the Leica renamer GUI.\cr
\code{\link{rename_zeiss_gui}} for the Zeiss renamer GUI.\cr
}
