% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeiss.r
\name{rename_zeiss}
\alias{rename_zeiss}
\title{Zeiss renamer}
\usage{
rename_zeiss(
  sourcefolder,
  targetfolder,
  infilepath,
  numrow,
  numcol,
  REGEXvalidimages = "^.*?s.*?t.*?\\\\.tif$",
  REGEXscenenum = "^.*/.*_s(.*)t.*\\\\.tif$",
  REGEXtimenum = "^.*/.*_s.*t(.*)\\\\.tif$",
  printMessages = TRUE,
  printFiles = TRUE,
  printToGUI = FALSE,
  move = FALSE
)
}
\arguments{
\item{sourcefolder}{character, the folder with raw images. The folder usually has the folllowing name structure: "demoplate_01--2019_01_01_12_00_00"}

\item{targetfolder}{character, folder where to place renamed files. A subfolder wih the same name as the infile will be created by the function and must not exist previously.}

\item{infilepath}{character, location of the microscope infile}

\item{numrow}{integer, number of rows in the multi well plate}

\item{numcol}{integer, number of columns in the multi well plate}

\item{REGEXvalidimages}{character, regular expression matching all files which should be renamed}

\item{REGEXscenenum}{character, regular expression capturing (\code{(\1)}) the scene number (i.e. well number)}

\item{REGEXtimenum}{character, regular expression capturing (\code{(\1)}) the time number in a time lapse}

\item{printMessages, }{logical, print messages in the console?}

\item{printFiles, }{logical, print file-by-file progress in the console?}

\item{printToGUI}{logical, print messages to the renamer GUI?}

\item{move}{logical, move (rather than copy) raw images?}
}
\value{
Renamed files, metadata and log file

Null
}
\description{
Renaming of Zeiss TIF files exported with Zen Blue. Files areLeica MatrixScreener files according to a microscope infile. See \code{\link{newinfile.df}} or \code{\link{newinfile.char}} for a description of the infile structure.
}
\seealso{
\code{\link{rename_leica_gui}} is the function which implements the actual renaming algorithm.\cr
}
