% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeiss.r
\name{rename_zeiss_gui}
\alias{rename_zeiss_gui}
\title{Zeiss renamer GUI}
\usage{
rename_zeiss_gui()
}
\value{
Null
}
\description{
Displays the renamer tool GUI for Zeiss Zen Blue exports.
The actual renaming process is performed by function rename_zeiss().
}
\seealso{
\code{\link{rename_zeiss}} is the function which implements the actual renaming algorithm.\cr
\code{\link{rename_leica_gui}} for the Leica renamer GUI.\cr
}
