% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{setup.dependency}
\alias{setup.dependency}
\title{Install and load dependencies}
\usage{
setup.dependency(pckgname)
}
\arguments{
\item{pckgname}{character, package name.}
}
\value{
\code{TRUE} if package is loaded successfully. \code{FALSE} otherwise.
}
\description{
Makes sure the target package is both installed and loaded into memory. Avoids unnecessary re-installs and re-loadings
}
