% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leica.r
\name{whichLeicaSoftware}
\alias{whichLeicaSoftware}
\title{Image export software}
\usage{
whichLeicaSoftware(folder)
}
\arguments{
\item{folder}{character, Leica export folder.}
}
\value{
character, the name of the software which generated the images.
}
\description{
Determines whether images coming from a Leica microscope were generated with the MatrixScreener or Navigator software.
}
